Weekly Logs Fully Implemented
+ Can now log to one file for one week. This means that a file will be
created every monday morning for the new week. If the program
crashes/gets closed we can restart the program and select "Start Day" in
admin window. This will allow us to select an existing csv file to write
to.
+ Removed number_records stat since unneeded
+ If a customer selects "Rental" then it will take them to the summary
page, skipping the problem page.
+ fixed bug in description text box where a user could press backspace
or space and it would not auto delete default text
+ Changed start/end time to 7:30/5:30 respectively
+ Changed time format in csv file to show date and military time